# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web Agent Bridge (WAB) API",
    "description": "Open protocol API for AI agent ↔ website interaction. Provides discovery, command execution, fairness-weighted search, and plan management for the Agentic Web.",
    "version": "1.2.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "WAB Support",
      "email": "support@webagentbridge.com",
      "url": "https://webagentbridge.com"
    },
    "termsOfService": "https://webagentbridge.com/terms",
    "x-logo": {
      "url": "https://webagentbridge.com/favicon.ico"
    }
  },
  "servers": [
    {
      "url": "https://webagentbridge.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Create account",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string", "minLength": 8},
                  "name": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Account created"},
          "400": {"description": "Validation error"}
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign in",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "JWT token returned"},
          "401": {"description": "Invalid credentials"}
        }
      }
    },
...
    "/api/noscript/bridge/{siteId}": {
      "get": {
        "tags": ["NoScript"],
        "summary": "NoScript bridge page",
        "description": "HTML bridge page for agents that cannot execute JavaScript. Provides structured page metadata via HTTP.",
        "operationId": "getNoscriptBridge",
        "parameters": [
          {"name": "siteId", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "HTML page with structured WAB metadata",
            "content": {"text/html": {}}
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "WAB session token obtained from /api/wab/authenticate"
      }
    },
...
  }
}